Spoke with you about changes been made.
For now needs to fix the dependencies part, instead of other python program we can manage to do it for each solution in the server side and send to client a json look like this:
array which each instance is json that look like this:
[ {solution: [course objects],dependencies: [ {courseNum : [dependencies] } , {courseNum : [dependencies] .... ] and so on.

need to make our algorithm more specific based on other preferences such as unwanted days and so on and you can look on the way i found more solutions it need to be changed also to something smarter but for now it is looking very very good.

you have index&conflicts which updates the file updated_courses.json based on conflicts and indexing. i changed main.py because found out that logic wasnt right related to conflicts so its good we found out.